requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
PyYAML==6.0.2
redis>=4.5.0
requests==2.32.5
urllib3==2.5.0

# Requirements for hyperparameter tuning
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.24.0
pandas>=1.1.0
matplotlib>=3.3.0
